+++ SpiritAnimalBackend/Models/SpiritAnimal.cs
@@ -2,13 +2,13 @@
 {
     public class SpiritAnimal
     {
-        public SpiritAnimal(long id, string name, string colour)
+        public SpiritAnimal(int id, string name, string colour)
         {
             Id = id;
             Name = name;
             Colour = colour;
         }
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Name { get; set; }
         public string? Colour { get; set; }
